Support #27540 - Optimize page query network requests
- Moved all QueryPage and QueryBuilder under a new component folder "list-page" to be organize the components.
- Create a new typescript file for storing all types (except props). To make it easier to find and use types across these components.
