+++ packages/common-ui/lib/index.ts
@@ -85,8 +85,8 @@ export * from "./formik-connected/NumberSpinnerField";
 export * from "./formik-connected/SelectFieldWithNav";
 export * from "./text-field-with-multiplication-button/TextFieldWithMultiplicationButton";
 export * from "./formik-connected/PlaceSectionsSelectionField";
-export * from "./query-builder/QueryBuilder";
+export * from "./list-page/QueryBuilder";
 export * from "./formik-connected/QueryLogicSwitchField";
-export * from "./query-page/QueryPage";
+export * from "./list-page/QueryPage";
 export * from "./formik-connected/ControlledVocabularySelectField";
 export * from "./formik-connected/FieldExtensionSelectField";
