--- packages/common-ui/lib/query-page/SearchResultsTable.tsx
@@ -1,3 +1,0 @@
-export function MultiSortTooltip() {
-  return;
-}
